requirements.txt
python-dotenv
langchain
langchain-google-genai
langchain-community
faiss-cpu
sentence-transformers
PyPDF2
langchain-text-splitters
streamlit